+++ app/gui2/src/components/visualizations/WarningsVisualization.vue
@@ -1,27 +1,22 @@
 <script lang="ts">
 export const name = 'Warnings'
 export const inputType = 'Any'
+export const defaultPreprocessor = [
+  'Standard.Visualization.Warnings',
+  'process_to_json_text',
+] as const
 </script>
 
 <script setup lang="ts">
-import { onMounted } from 'vue'
-
 import VisualizationContainer from '@/components/VisualizationContainer.vue'
 
 type Data = string[]
 
 const props = defineProps<{ data: Data }>()
-const emit = defineEmits<{
-  'update:preprocessor': [module: string, method: string, ...args: string[]]
-}>()
-
-onMounted(() => {
-  emit('update:preprocessor', 'Standard.Visualization.Warnings', 'process_to_json_text')
-})
 </script>
 
 <template>
-  <VisualizationContainer :below-toolbar="true">
+  <VisualizationContainer :belowToolbar="true">
     <div class="WarningsVisualization">
       <ul>
         <li v-if="props.data.length === 0">There are no warnings.</li>