+++ lib/rust/metamodel/src/lib.rs
@@ -45,6 +45,7 @@
 
 // === Features ===
 #![feature(option_get_or_insert_default)]
+#![feature(type_alias_impl_trait)]
 // === Standard Linter Configuration ===
 #![deny(non_ascii_idents)]
 #![warn(unsafe_code)]